+++ doc/changelog.md
@@ -1,6 +1,11 @@
 
 # Change Log
 
+## 0.9.2 (2013-06-03)
+
+- Add "json", "array" as type for field
+- Add "nullfill" option
+
 ## 0.9.1
 
 - Add development mode to ignore cache and not to save cache